Add page layout styling
Updated the CSS stylesheet to provide the overall look of the page.
Aside from styling the labels, no form styling has been applied yet.
This will be implemented in the following commit.

To make the form layout styling more straightforward, modified the html
to separate out the input and labels for the form. Previously, the input
had been nested inside of the label. This made repositioning the label,
and creating the layout for the form, trickier. Separating these out and
forcing them to be block elements simplified this process.