Add form input styling
Added basic styling to form input by changing the focus border color and
adding a red border color if passwords are invalid.

Modified the html to make passwords required as a way to trigger the
invalid response and display the change to the border color. Further
added an empty bold element under password which can be used to display
a message when the password input is invalid.